demo version of DSSAN

+++ app/Http/Controllers/UserController.php
@@ -13,23 +13,25 @@ class UserController extends Controller
 {
     public function index()
     {
-        $users = User::orderBy('role', 'desc')->get();
-        return view('admin.user.index', compact('users'));
+        $users = User::orderBy("role", "desc")->get();
+        return view("admin.user.index", compact("users"));
     }
 
     public function create()
     {
-        return view('admin.user.create');
+        return view("admin.user.create");
     }
 
     public function store(UserFormRequest $request)
     {
         $validatedData = $request->validated();
-        $validatedData['image'] = $request->file('image')->store('userImage');
-        $validatedData['password'] = Hash::make('password');
+        $validatedData["image"] = $request->file("image")->store("userImage");
+        $validatedData["password"] = Hash::make("password");
         User::create($validatedData);
-        Password::sendResetLink($request->only(['email']));
-        return redirect()->route('user.index')->with('success', 'User added successfully.');
+        Password::sendResetLink($request->only(["email"]));
+        return redirect()
+            ->route("user.index")
+            ->with("success", "User added successfully.");
     }
 
     public function show($id)
@@ -39,24 +41,32 @@ class UserController extends Controller
 
     public function edit(User $user)
     {
-        return view('admin.user.edit', compact('user'));
+        return view("admin.user.edit", compact("user"));
     }
 
     public function update(UserFormRequest $request, User $user)
     {
         $validatedData = $request->validated();
-        if ($request->hasFile('image')) {
-            $validatedData['image'] = $request->file('image')->store('userImage');
+        if ($request->hasFile("image")) {
+            $validatedData["image"] = $request
+                ->file("image")
+                ->store("userImage");
             !is_null($user->image) ?? Storage::delete($user->image);
         }
         $user->update($validatedData);
-        return redirect()->route('user.index')->with('info', 'Details updated successfully.');
+        return redirect()
+            ->route("user.index")
+            ->with("info", "Details updated successfully.");
     }
 
     public function destroy(User $user)
     {
-        Storage::delete($user->image);
+        if (!is_null($user->image)) {
+            Storage::delete($user->image);
+        }
         $user->delete();
-        return redirect()->route('user.index')->with('danger', 'User removed successfully.');
+        return redirect()
+            ->route("user.index")
+            ->with("danger", "User removed successfully.");
     }
 }
